airflow/requirements.txt
# Airflow Requirements for NeuroNews Custom Image (Issue #187)
# This file contains all Python dependencies needed for our Airflow environment
# including OpenLineage providers and DAG dependencies.

# Core Airflow extras and providers
apache-airflow[postgres,celery,redis,ssh,ftp,sftp,s3,docker,kubernetes]==2.8.1

# OpenLineage providers for data lineage tracking
apache-airflow-providers-openlineage==1.4.0
openlineage-airflow==1.9.0
openlineage-python==1.9.0

# Database providers
apache-airflow-providers-postgres==5.8.0
apache-airflow-providers-mysql==5.4.0

# Cloud providers for AWS integration
apache-airflow-providers-amazon==8.15.0
apache-airflow-providers-http==4.7.0

# Data processing and analysis libraries (for DAGs)
pandas==2.1.4
numpy==1.24.4
requests==2.31.0
sqlalchemy==2.0.25

# News scraping and processing dependencies
scrapy==2.11.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
feedparser==6.0.10

# Machine learning and NLP libraries
scikit-learn==1.3.2
nltk==3.8.1
textblob==0.17.1

# Graph database integration
gremlinpython==3.7.0

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# Data validation and quality
great-expectations==0.18.8
pandera==0.17.2

# JSON and data serialization
orjson==3.9.10
pydantic==2.5.2

# HTTP clients and API integration
httpx==0.25.2
aiohttp==3.9.1

# Date and time utilities
python-dateutil==2.8.2
pytz==2023.3

# Configuration management
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing and development (optional for runtime)
pytest==7.4.3
pytest-asyncio==0.21.1

# Security and encryption
cryptography==41.0.8
bcrypt==4.1.2

# Task queue and messaging
celery==5.3.4
redis==5.0.1

# File format support
openpyxl==3.1.2
xlrd==2.0.1

# Network and connectivity
psycopg2-binary==2.9.9
boto3==1.34.0
